add buttons and stat pages.

+++ force_commands.py
@@ -78,7 +78,7 @@ class forceModal(Modal):
             "Link": googledoc,
             "Leader": leader.id,
             "Desc": desc,
-            "Image": " ",
+            "Image": image,
             "Colour": colour.strip('#'),
             "Ranking": 0,
             "MemberCount": 1,
@@ -88,15 +88,19 @@ class forceModal(Modal):
 
         if self.edit:
             forces = db["Forces"] 
+            forceFound=False
             for i in forces:
                 if i['Name'] == self.oldName:
+                    forceFound == True
                     index = forces.index(i)
                     role = interaction.guild.get_role(self.oldRole)
-                    role = await role.edit(name=name,color=int(colour.strip('#'),16))
+                    await role.edit(name=name,color=int(colour.strip('#'),16))
                     newForce['RoleID'] = role.id
                     forces[index] = newForce
                     embed = createEmbed(newForce,leader)
-                    await interaction.response.send_message(f"{name} has been edited!", embed=embed,ephemeral=True)             
+                    await interaction.response.send_message(f"{name} has been edited!", embed=embed,ephemeral=True)
+            if forceFound == False:
+                await interaction.response.send_message(f"force {name} cannot be found.", embed=embed,ephemeral=True)
         else:
             if validation.doesKeyExist("Forces"):
                 forces = db["Forces"] 
@@ -131,64 +135,6 @@ class ForceCommands(commands.Cog):
         print('command create force activated')
         modal = forceModal(title=f"Create a Force ")
         await ctx.send_modal(modal)
-    
-#    @force.command(guild_ids=[*guildids], description='Create a Force')
-#   async def create(self, ctx,
-#       name: Option(str,"the name of the force", required=True),
-#       googledoc: Option(str, "the link to the forces google doc"),
-#       colour: Option(str, "the colour associated with the force in hexademical form ex: ffffff would be white", default='ffffff'),
-#        leader: Option(discord.Member, "the leader of the force leave blank if it is you", required=False,default=None)
-#    ):
-#        if leader == None:
-#            leader = ctx.author
-#        if not validation.validHexColor(colour):
-#            await ctx.respond("invalid colour given please use colours in hex value ex: #19778a")
-#            return
-#        if validation.doesKeyExist("Forces"):
-#            
-#           forces = db["Forces"] 
-#            if checkDupeName(name, forces):
-#                await ctx.respond(f'A force with the name {name} already exists please chose another name.', ephemeral=True)
-#            else:
-#               forceDict ={
-#                    "Name": name,
-#                    "Link": googledoc,
-#                    "Leader": leader.id,
-#                    "Desc": " ",
-#                    "Image": " ",
-#                    "Colour": colour.strip('#'),
-#                    "Ranking": 0,
-#                    "Members": [],
-#                    "MemberCount": 1,
-#                    "RoleID": 0
-#                    
-#                }
-#                role = await ctx.guild.create_role(name=name, color=int(colour.strip('#'),16),reason="new force role")
-#                await leader.add_roles(role)
-#                forceDict["RoleID"] = role.id
-#                forces.append(forceDict)
-#                db["Forces"] = forces
-#                await ctx.respond(f"{name} has been created!")
-#            
-#        else:
-#            newForce = {
-#                "Name": name,
-#                "Link": googledoc,
-#                "Leader": leader.id,
-#                "Desc": " ",
-#                "Image": " ",
-#                "Colour": colour.strip('#'),
-#                "Ranking": 0,
-#                "MemberCount": 1,
-#                "Members": [],
-#                "RoleID": 0
-#            }
-#
-#            role = await ctx.guild.create_role(name=name, color=int(colour,16),reason="new force role")
-#            await leader.add_roles(role)
-#            newForce["RoleID"] = role.id
-#            db["Forces"] = [newForce]
-#            await ctx.respond(f"{name} has been created!")
 
     @force.command(guild_ids=[*guildids], description="get the info for a force")
     async def get(
@@ -216,46 +162,18 @@ class ForceCommands(commands.Cog):
     @force.command(guild_ids=[*guildids], description="edit the info of a force")
     async def edit(
         self, ctx, 
-        force: Option(str,'the force you want to get. ex: \'build divers\' or \'pizza grubbers\'' ,required=True),
-        inputfield: Option(str,'the data of the force you want to change',choices=["Name","Link","Desc","Image","Leader","Colour"], required=True),
-        inputdata: Option(str,'the new data for the inputfield, like the new name or color', required=True),
+        force: Option(str,'the force you want to edit. ex: \'build divers\' or \'pizza grubbers\'' ,required=True),
         public: Option(bool, "makes the message only visible to you if false, True by default",required=False, default=False)
     ):
-        if validation.doesKeyExist("Forces"):
-            forces = db["Forces"]
-            forceFound = False
-            for i in forces:
-                if i['Name'].casefold().startswith(force.casefold()):
-                    if i['Leader'] == ctx.author.id or validation.userHasRole(ctx.author, adminRoles):
-                        guild = ctx.guild
-                        role = guild.get_role(i['RoleID'])
-                        if inputfield == 'Leader':
-                            inputdata = int(inputdata)
-                            try:  member= await ctx.guild.fetch_member(inputdata)
-                            except: 
-                                await ctx.respond('please input a valid user id', ephemeral=not public)
-                                return
-                            await member.add_roles(role)
-                        elif inputfield == 'Colour':
-                            if validation.validHexColor(inputdata):
-                                 await role.edit(colour=int(inputdata.strip('#'),16))
-                            else:
-                                await ctx.respond("invalid colour given please use colours in hex value ex: #19778a", ephemeral=not public)
-                                return
-                        elif inputfield == 'Name':
-                            await role.edit(name=inputdata)
-                        index= forces.index(i)
-                        forces[index][inputfield] = inputdata
-                        await ctx.respond(f'{inputfield} changed to {inputdata} on {i["Name"]}',ephemeral=not public)
-                    else:
-                        ctx.respond("You do not have permission to edit someone else's forms.",ephemeral=True)
-                        return
-                    forceFound = True
-                    break
-            if not forceFound:
-                await ctx.respond(f'no force found with the name: {force}', ephemeral=True)
-        else: 
-            await ctx.respond(f'no force found with the name: {force}', ephemeral=True)
+        forces = db['Forces'] 
+        for i in forces:
+            if i['Name'].casefold().startswith(force.casefold()):
+                if ctx.author.id == i['Leader']:
+                    modal = forceModal(title=f"Edit {i['Name']}.",edit=True,oldValues=i)
+                    await ctx.send_modal(modal)
+                else:
+                    ctx.respond('you are not the leader of the force',ephemeral=True)
+    
     @force.command(guild_ids=[*guildids], description="delete a force")
     async def delete(self, ctx, 
                      force: Option(str,'the force you want to get. ex: \'build divers\' or \'pizza grubbers\'' ,required=True),
@@ -268,7 +186,7 @@ class ForceCommands(commands.Cog):
                 if i['Name'].casefold().startswith(force.casefold()):
                     if i['Leader'] == ctx.author.id or validation.userHasRole(ctx.author, adminRoles):
                         role = ctx.guild.get_role(i['RoleID'])
-                        await role.delete()
+                        if role != None: await role.delete()
                         forces.remove(i)
                         await ctx.respond(f"force deleted {force}")
                     else: await ctx.respond("you do not have permision to delete someone else's force", ephemeral= not public)
